+++ internal/domain/dto/taskdto/create.go
@@ -5,7 +5,7 @@ import (
 )
 
 type CreateRequest struct {
-	UserID      uint              `json:"user_id"`
+	UserID      uint64            `json:"user_id"`
 	Title       string            `json:"title"`
 	Description string            `json:"description"`
 	Status      entity.TaskStatus `json:"status"`
