+++ internal/domain/dto/userdto/tasks_info.go
@@ -1,7 +1,7 @@
 package userdto
 
 type TasksRequest struct {
-	ID uint `param:"id" json:"id"`
+	ID uint64 `param:"id" json:"id"`
 }
 
 type TasksResponse struct {
